Change constraint for jQuery and bump up vis-ui
Digitier is not able to initialize with the searchbar in the current release. We updated jQuery to 1.11.3 in a prior commit . 1.11.3 contains a bugfix which will break the used dataTables version. So we update jQuery to a newer version and also bump up the datatables version to the newest release in the vis-ui package. There are some other libs with newer versions too  because of composer update